move azure_kinect
