+++ cpp/open3d/io/sensor/azure_kinect/AzureKinectSensor.cpp
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/io/sensor/AzureKinect/AzureKinectSensor.h"
+#include "open3d/io/sensor/azure_kinect/AzureKinectSensor.h"
 
 #include <k4a/k4a.h>
 #include <k4arecord/record.h>
@@ -32,7 +32,7 @@
 #include <memory>
 
 #include "open3d/geometry/RGBDImage.h"
-#include "open3d/io/sensor/AzureKinect/K4aPlugin.h"
+#include "open3d/io/sensor/azure_kinect/K4aPlugin.h"
 
 namespace open3d {
 namespace io {
