+++ cpp/open3d/io/sensor/azure_kinect/AzureKinectSensorConfig.h
