+++ cpp/open3d/io/sensor/azure_kinect/MKVMetadata.cpp
@@ -24,58 +24,39 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#pragma once
+#include "open3d/io/sensor/azure_kinect/MKVMetadata.h"
 
-#include <memory>
-
-#include "open3d/io/sensor/AzureKinect/AzureKinectSensorConfig.h"
-#include "open3d/io/sensor/RGBDSensor.h"
-
-struct _k4a_capture_t;         // typedef _k4a_capture_t* k4a_capture_t;
-struct _k4a_device_t;          // typedef _k4a_device_t* k4a_device_t;
-struct _k4a_transformation_t;  // typedef _k4a_transformation_t*
-                               // k4a_transformation_t;
+#include <json/json.h>
+#include <unordered_map>
 
 namespace open3d {
-namespace geometry {
-class RGBDImage;
-class Image;
-}  // namespace geometry
-
 namespace io {
 
-// Avoid including AzureKinectRecorder.h
-class AzureKinectRecorder;
+bool MKVMetadata::ConvertToJsonValue(Json::Value &value) const {
+    intrinsics_.ConvertToJsonValue(value);
 
-/// \class AzureKinectSensor
-///
-/// AzureKinect sensor.
-class AzureKinectSensor : public RGBDSensor {
-public:
-    /// \brief Default Constructor.
-    AzureKinectSensor(const AzureKinectSensorConfig& sensor_config);
-    ~AzureKinectSensor();
+    value["serial_number_"] = serial_number_;
+    value["color_mode"] = color_mode_;
+    value["depth_mode"] = depth_mode_;
 
-    bool Connect(size_t sensor_index) override;
-    std::shared_ptr<geometry::RGBDImage> CaptureFrame(
-            bool enable_align_depth_to_color) const override;
+    value["stream_length_usec"] = stream_length_usec_;
+    value["width"] = width_;
+    value["height"] = height_;
 
-    static bool PrintFirmware(_k4a_device_t* device);
-    /// List available Azure Kinect devices.
-    static bool ListDevices();
-    static std::shared_ptr<geometry::RGBDImage> DecompressCapture(
-            _k4a_capture_t* capture, _k4a_transformation_t* transformation);
+    return true;
+}
+bool MKVMetadata::ConvertFromJsonValue(const Json::Value &value) {
+    intrinsics_.ConvertFromJsonValue(value);
 
-protected:
-    _k4a_capture_t* CaptureRawFrame() const;
+    serial_number_ = value["serial_number"].asString();
+    color_mode_ = value["color_mode"].asString();
+    depth_mode_ = value["depth_mode"].asString();
 
-    AzureKinectSensorConfig sensor_config_;
-    _k4a_transformation_t* transform_depth_to_color_;
-    _k4a_device_t* device_;
-    int timeout_;
+    stream_length_usec_ = value["stream_length_usec"].asUInt64();
+    width_ = value["width"].asInt();
+    height_ = value["height"].asInt();
 
-    friend class AzureKinectRecorder;
-};
-
+    return true;
+}
 }  // namespace io
 }  // namespace open3d
