+++ cpp/open3d/io/sensor/azure_kinect/MKVReader.h
@@ -24,39 +24,56 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/io/sensor/AzureKinect/MKVMetadata.h"
+#pragma once
 
-#include <json/json.h>
-#include <unordered_map>
+#include "open3d/geometry/RGBDImage.h"
+#include "open3d/io/sensor/azure_kinect/MKVMetadata.h"
+#include "open3d/utility/IJsonConvertible.h"
+
+struct _k4a_playback_t;        // typedef _k4a_playback_t* k4a_playback_t;
+struct _k4a_capture_t;         // typedef _k4a_capture_t* k4a_capture_t;
+struct _k4a_transformation_t;  // typedef _k4a_transformation_t*
+                               // k4a_transformation_t;
 
 namespace open3d {
 namespace io {
 
-bool MKVMetadata::ConvertToJsonValue(Json::Value &value) const {
-    intrinsics_.ConvertToJsonValue(value);
+/// \class MKVReader
+///
+/// AzureKinect mkv file reader.
+class MKVReader {
+public:
+    /// \brief Default Constructor.
+    MKVReader();
+    virtual ~MKVReader() {}
 
-    value["serial_number_"] = serial_number_;
-    value["color_mode"] = color_mode_;
-    value["depth_mode"] = depth_mode_;
+    /// Check If the mkv file is opened.
+    bool IsOpened();
+    /// Check if the mkv file is all read.
+    bool IsEOF() { return is_eof_; }
 
-    value["stream_length_usec"] = stream_length_usec_;
-    value["width"] = width_;
-    value["height"] = height_;
+    /// Open an mkv playback.
+    ///
+    /// \param filename Path to the mkv file.
+    bool Open(const std::string &filename);
+    /// Close the opened mkv playback.
+    void Close();
 
-    return true;
-}
-bool MKVMetadata::ConvertFromJsonValue(const Json::Value &value) {
-    intrinsics_.ConvertFromJsonValue(value);
+    /// Get metadata of the mkv playback.
+    MKVMetadata &GetMetadata() { return metadata_; }
+    /// Seek to the timestamp (in us).
+    bool SeekTimestamp(size_t timestamp);
+    /// Get next frame from the mkv playback and returns the RGBD object.
+    std::shared_ptr<geometry::RGBDImage> NextFrame();
 
-    serial_number_ = value["serial_number"].asString();
-    color_mode_ = value["color_mode"].asString();
-    depth_mode_ = value["depth_mode"].asString();
+private:
+    _k4a_playback_t *handle_;
+    _k4a_transformation_t *transformation_;
+    MKVMetadata metadata_;
+    bool is_eof_ = false;
 
-    stream_length_usec_ = value["stream_length_usec"].asUInt64();
-    width_ = value["width"].asInt();
-    height_ = value["height"].asInt();
-
-    return true;
-}
+    Json::Value GetMetadataJson();
+    std::string GetTagInMetadata(const std::string &tag_name);
+};
 }  // namespace io
 }  // namespace open3d
